update user authorization

+++ src/api/notes/handler.js
@@ -2,18 +2,18 @@ class NotesHandler {
   constructor(service, validator) {
     this._service = service;
     this._validator = validator;
-
-    this.postNoteHandler = this.postNoteHandler.bind(this);
-    this.getNotesHandler = this.getNotesHandler.bind(this);
-    this.getNoteByIdHandler = this.getNoteByIdHandler.bind(this);
-    this.putNoteByIdHandler = this.putNoteByIdHandler.bind(this);
-    this.deleteNoteByIdHandler = this.deleteNoteByIdHandler.bind(this);
   }
 
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
-    const noteId = await this._service.addNote({ title, body, tags });
+
+    // authorization
+    const { id: credentialId } = request.auth.credentials;
+
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialId,
+    });
 
     return h.response({
       status: 'success',
@@ -22,16 +22,24 @@ class NotesHandler {
     }).code(201);
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
-    return {
+  async getNotesHandler(request, h) {
+    // authorization
+    const { id: credentialId } = request.auth.credentials;
+
+    const notes = await this._service.getNotes(credentialId);
+    return h.response({
       status: 'success',
       data: { notes },
-    };
+    });
   }
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params; // ambil nilai id dari path parameter
+
+    // authorization
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
+
     // masukkan ke var baru menggunakan NotesService
     const note = await this._service.getNoteById(id);
     return h.response({
@@ -43,6 +51,11 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+
+    // authorization
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
+
     await this._service.editNoteById(id, request.payload);
     return h.response({
       status: 'success',
@@ -52,6 +65,11 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+
+    // authorization
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
+
     await this._service.deleteNoteById(id);
     return h.response({
       status: 'success',
